+++ tables.h
@@ -71,6 +71,7 @@ extern const angle_t tantoangle[SLOPERANGE+1];
 
 // Utility function, called by R_PointToAngle.
 int SlopeDiv(unsigned num, unsigned den);
+int SlopeDivCrispy(unsigned num, unsigned den);   // overflow-safe (large maps)
 int arctan(int x, int y); //sf
 
 #endif